requirements.txt
Django>=4.2,<7
djangorestframework>=3.14
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
